// src/modules/academicFaculty/academicFaculty.interface.ts
export type TAcademicFaculty = {
    name: string;
};

export interface FreedomFormData {
  name: string;
  date: string;

  desiredAnnualIncome: string;
  expectedNonInvestmentIncome: string;

  businessOwnershipOutlook: string;
  personalDebts: string;

  professionalFees: string;
  employeeRetentionBonuses: string;
  outstandingDebtOnHand: string;

  taxOnProceeds: string;
}